extract randomness
